个人博客0.2

+++ 后端/vueblog/src/main/java/world/lixiang/controller/TagController.java
@@ -5,8 +5,11 @@ import org.springframework.web.bind.annotation.*;
 import world.lixiang.entity.Tag;
 import world.lixiang.entity.Type;
 import world.lixiang.service.TagService;
+import world.lixiang.utils.JwtInfo;
+import world.lixiang.utils.JwtUtils;
 import world.lixiang.vo.Result;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -53,14 +56,20 @@ public class TagController {
     }
 
     @GetMapping("deleteTag")
-    public Result deleteTag(Integer id){
+    public Result deleteTag(Integer id , HttpServletRequest request){
         Result result = new Result();
         try{
+            JwtInfo jwtInfo = JwtUtils.getMemberIdByJwtToken(request);
+            String id1 = jwtInfo.getId();
+            if(!id1.equals("1")){
+                throw new RuntimeException("权限不够");
+            }
             tagService.deleteTag(id);
             result = result.success("删除成功");
-        }catch (Exception e){
-            result = result.error("删除失败");
+        }catch (RuntimeException e){
             e.printStackTrace();
+            result =  result.error("权限不够");;
+
         }
         return  result;
     }
